clang tidy

+++ src/protocol/kademlia/message.cpp
@@ -46,7 +46,8 @@ namespace libp2p::protocol::kademlia {
 
       auto peer_id_res = PeerId::fromBytes(gsl::span<const uint8_t>(
           // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-          reinterpret_cast<const uint8_t *>(src.id().data()), src.id().size()));
+          reinterpret_cast<const uint8_t *>(src.id().data()),
+          gsl::narrow<ptrdiff_t>(src.id().size())));
       if (!peer_id_res) {
         return Message::Error::INVALID_PEER_ID;
       }
@@ -55,7 +56,8 @@ namespace libp2p::protocol::kademlia {
       for (const auto &addr : src.addrs()) {
         auto res = multi::Multiaddress::create(gsl::span<const uint8_t>(
             // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-            reinterpret_cast<const uint8_t *>(addr.data()), addr.size()));
+            reinterpret_cast<const uint8_t *>(addr.data()),
+            gsl::narrow<ptrdiff_t>(addr.size())));
         if (!res) {
           return Message::Error::INVALID_ADDRESSES;
         }
@@ -105,7 +107,7 @@ namespace libp2p::protocol::kademlia {
   bool Message::deserialize(const void *data, size_t sz) {
     clear();
     pb::Message pb_msg;
-    if (!pb_msg.ParseFromArray(data, sz)) {
+    if (!pb_msg.ParseFromArray(data, gsl::narrow<int>(sz))) {
       error_message_ = "Invalid protobuf data";
       return false;
     }
@@ -175,7 +177,7 @@ namespace libp2p::protocol::kademlia {
     buffer.resize(prefix_sz + msg_sz);
     memcpy(buffer.data(), varint_vec.data(), prefix_sz);
     return pb_msg.SerializeToArray(buffer.data() + prefix_sz,  // NOLINT
-                                   msg_sz);
+                                   gsl::narrow<int>(msg_sz));
   }
 
   void Message::selfAnnounce(PeerInfo self) {
